"Updated Jenkinsfile to enhance CI/CD pipeline"
Upgraded Jenkinsfile to ensure a more robust continuous integration and deployment process. Environment variables were added for increased adaptability. Created stages for initiating the build process, running tests and notifying the start and end of the build process. Made the build process more dynamic by depending on the branch name. The Dockerfile was also made adaptable depending on the branch. Lastly, added feature for remote update in the deployment stage.