"Removed unnecessary whitespace from Jenkinsfile"
Whitespace on line 203 in the Jenkinsfile was eliminated. This is purely a cosmetic change aimed at maintaining code cleanliness and readability. No functionality of the file has been altered. The removal of excess and unnecessary spaces makes the Jenkinsfile look neater, making the stages of the build process more distinguishable and easier to follow.