Improved Docker update process in Jenkinsfile
Replaced the process of updating Docker service on DevelopHost which was initially handled by withCredentials and sshCommand. This change simplifies Docker service handling since the service update command is directly implemented within the Jenkins node. This results in removing the need for ssh credentials which makes the process more efficient.